+++ packages/app/src/App.tsx
@@ -45,14 +45,14 @@ import {
   ScaffolderPage,
   scaffolderPlugin,
   ScaffolderFieldExtensions,
-  RepoUrlPickerFieldExtension,
-  OwnerPickerFieldExtension,
-  EntityPickerFieldExtension,
-  EntityNamePickerFieldExtension,
 } from '@backstage/plugin-scaffolder';
 import { SearchPage } from '@backstage/plugin-search';
 import { TechRadarPage } from '@backstage/plugin-tech-radar';
-import { TechdocsPage } from '@backstage/plugin-techdocs';
+import {
+  DefaultTechDocsHome,
+  TechDocsIndexPage,
+  TechDocsReaderPage,
+} from '@backstage/plugin-techdocs';
 import { UserSettingsPage } from '@backstage/plugin-user-settings';
 import AlarmIcon from '@material-ui/icons/Alarm';
 import React from 'react';
@@ -116,13 +116,15 @@ const routes = (
       {entityPage}
     </Route>
     <Route path="/catalog-import" element={<CatalogImportPage />} />
-    <Route path="/docs" element={<TechdocsPage />} />
+    <Route path="/docs" element={<TechDocsIndexPage />}>
+      <DefaultTechDocsHome />
+    </Route>
+    <Route
+      path="/docs/:namespace/:kind/:name/*"
+      element={<TechDocsReaderPage />}
+    />
     <Route path="/create" element={<ScaffolderPage />}>
       <ScaffolderFieldExtensions>
-        <EntityPickerFieldExtension />
-        <EntityNamePickerFieldExtension />
-        <RepoUrlPickerFieldExtension />
-        <OwnerPickerFieldExtension />
         <LowerCaseValuePickerFieldExtension />
       </ScaffolderFieldExtensions>
     </Route>